MODPATBLK-13 Refactor tests
