+++ src/test/java/org/folio/rest/handlers/ItemDeclaredLostEventHandlerTest.java
@@ -8,9 +8,7 @@ import java.util.List;
 
 import org.folio.domain.OpenLoan;
 import org.folio.domain.UserSummary;
-import org.folio.repository.UserSummaryRepository;
 import org.folio.rest.jaxrs.model.ItemDeclaredLostEvent;
-import org.folio.rest.utils.UserSummaryUtils;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -19,11 +17,9 @@ import io.vertx.ext.unit.TestContext;
 import io.vertx.ext.unit.junit.VertxUnitRunner;
 
 @RunWith(VertxUnitRunner.class)
-public class ItemDeclaredLostEventHandlerTest extends UserSummaryUtils {
+public class ItemDeclaredLostEventHandlerTest extends EventHandlerTest {
   private static final ItemDeclaredLostEventHandler eventHandler =
     new ItemDeclaredLostEventHandler(postgresClient);
-  protected static final UserSummaryRepository userSummaryRepository =
-    new UserSummaryRepository(postgresClient);
 
   @Before
   public void beforeEach(TestContext context) {
@@ -47,7 +43,7 @@ public class ItemDeclaredLostEventHandlerTest extends UserSummaryUtils {
       .withNumberOfLostItems(1)
       .withOutstandingFeeFineBalance(BigDecimal.ZERO);
 
-    checkUserSummary(summaryId, userSummaryToCompare, context, userSummaryRepository);
+    checkUserSummary(summaryId, userSummaryToCompare, context);
   }
 
   @Test
@@ -73,7 +69,7 @@ public class ItemDeclaredLostEventHandlerTest extends UserSummaryUtils {
       .withNumberOfLostItems(1)
       .withOutstandingFeeFineBalance(BigDecimal.ZERO);
 
-    checkUserSummary(summaryId, userSummaryToCompare, context, userSummaryRepository);
+    checkUserSummary(summaryId, userSummaryToCompare, context);
   }
 
   @Test
@@ -105,6 +101,6 @@ public class ItemDeclaredLostEventHandlerTest extends UserSummaryUtils {
       .withNumberOfLostItems(1)
       .withOutstandingFeeFineBalance(BigDecimal.ZERO);
 
-    checkUserSummary(summaryId, userSummaryToCompare, context, userSummaryRepository);
+    checkUserSummary(summaryId, userSummaryToCompare, context);
   }
 }
